backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
nanoid==2.0.0
sentence-transformers==2.7.0
pinecone-client==3.2.2
numpy==1.26.4
python-multipart==0.0.9
httpx==0.27.0
